+++ TESTING.md
@@ -1,0 +1,3 @@
+# Testing
+
+1. Install and activate for the first time on a brand new WordPress install.
